requirements.txt
numpy==1.19.5
tensorflow_cpu==2.4.1
scikit_learn==0.24.1
